remove casts

+++ src/instructions.rs
@@ -34,40 +34,40 @@ fn jump_to_address(cpu: &mut Chip8, opcode: OpCode) {
 
 fn jump_if_reg_values_are_equal(cpu: &mut Chip8, opcode: OpCode) {
     let (vx, vy) = get_vx_and_vy(&opcode);
-    if cpu.registers[vx as usize] == cpu.registers[vy as usize] {
+    if cpu.registers[vx] == cpu.registers[vy] {
         cpu.pc += 2;
     }
 }
 
 fn jump_if_reg_value_is_equal_to_number(cpu: &mut Chip8, opcode: OpCode) {
     let (vx, number) = get_vx_and_number(&opcode);
-    if cpu.registers[vx as usize] == number as u8 {
+    if cpu.registers[vx] == number {
         cpu.pc += 2;
     }
 }
 
 fn jump_if_reg_value_is_not_equal_to_number(cpu: &mut Chip8, opcode: OpCode) {
     let (vx, number) = get_vx_and_number(&opcode);
-    if cpu.registers[vx as usize] != number as u8 {
+    if cpu.registers[vx] != number {
         cpu.pc += 2;
     }
 }
 
 fn jump_if_reg_values_are_not_equal(cpu: &mut Chip8, opcode: OpCode) {
     let (vx, vy) = get_vx_and_vy(&opcode);
-    if cpu.registers[vx as usize] != cpu.registers[vy as usize] {
+    if cpu.registers[vx] != cpu.registers[vy] {
         cpu.pc += 2;
     }
 }
 
 fn store_number_in_register(cpu: &mut Chip8, opcode: OpCode) {
     let (vx, number) = get_vx_and_number(&opcode);
-    cpu.registers[vx as usize] = number as u8;
+    cpu.registers[vx] = number;
 }
 
 fn copy_register_value(cpu: &mut Chip8, opcode: OpCode) {
     let (vx, vy) = get_vx_and_vy(&opcode);
-    cpu.registers[vx as usize] = cpu.registers[vy as usize];
+    cpu.registers[vx] = cpu.registers[vy];
 }
 
 fn add_to_register_and_ignore_carry_flag(_cpu: &mut Chip8, _opcode: OpCode) {
@@ -76,17 +76,17 @@ fn add_to_register_and_ignore_carry_flag(_cpu: &mut Chip8, _opcode: OpCode) {
 
 fn bitwise_or_and_store(cpu: &mut Chip8, opcode: OpCode) {
     let (vx, vy) = get_vx_and_vy(&opcode);
-    cpu.registers[vx as usize] |= cpu.registers[vy as usize];
+    cpu.registers[vx] |= cpu.registers[vy];
 }
 
 fn bitwise_and_and_store(cpu: &mut Chip8, opcode: OpCode) {
     let (vx, vy) = get_vx_and_vy(&opcode);
-    cpu.registers[vx as usize] &= cpu.registers[vy as usize];
+    cpu.registers[vx] &= cpu.registers[vy];
 }
 
 fn bitwise_xor_and_store(cpu: &mut Chip8, opcode: OpCode) {
     let (vx, vy) = get_vx_and_vy(&opcode);
-    cpu.registers[vx as usize] ^= cpu.registers[vy as usize];
+    cpu.registers[vx] ^= cpu.registers[vy];
 }
 
 fn bitwise_shif_right_and_store(_cpu: &mut Chip8, _opcode: OpCode) {
@@ -95,12 +95,12 @@ fn bitwise_shif_right_and_store(_cpu: &mut Chip8, _opcode: OpCode) {
 
 fn add_and_store(cpu: &mut Chip8, opcode: OpCode) {
     let (vx, vy) = get_vx_and_vy(&opcode);
-    cpu.registers[vx as usize] += cpu.registers[vy as usize];
+    cpu.registers[vx] += cpu.registers[vy];
 }
 
 fn subtract_and_store(cpu: &mut Chip8, opcode: OpCode) {
     let (vx, vy) = get_vx_and_vy(&opcode);
-    cpu.registers[vx as usize] -= cpu.registers[vy as usize];
+    cpu.registers[vx] -= cpu.registers[vy];
 }
 
 fn subtract_and_store_and_set_vf(_cpu: &mut Chip8, _opcode: OpCode) {
